+++ lib/src/flusbserial/cp210x_serial_device.dart
@@ -1,5 +1,6 @@
 import 'dart:ffi';
 import 'dart:typed_data';
+import 'dart:io';
 
 import 'package:ffi/ffi.dart' as ffi;
 import 'package:flusbserial/src/models/usb_configuration.dart';
@@ -87,6 +88,10 @@ class Cp210XSerialDevice extends UsbSerialDevice {
     }
 
     deviceHandle = handle;
+
+    if (UsbSerialDevice.autoDetachKernelDriverEnabled && Platform.isLinux) {
+      _libusb.libusb_set_auto_detach_kernel_driver(deviceHandle, 1);
+    }
 
     configuration = await getConfiguration(0);